change register position
